requirements.txt
face-recognition==0.1.11
google-cloud-pubsub==2.13.0
opencv-python==4.6.0.66
